time: add wasip1 files
Add wasip1 files to the time package.

When writing this, I found out the horrifying fact that wasip1 doesn't
specify what time zone the runtime is using. This will be corrected in
wasip2, but for now we can make an assumption that it's using UTC, as
that's what wasmtime, wasmer, and wazero use for clock_time_get on the
realtime clock.

The list of files to read from also doesn't make sense in wasip1
(there's no consistent platform filesystem layout, runtimes can and will
define their own filesystems with their own semantics), so it is left
with an empty slice to satisfy the compiler.

Change-Id: Ib1db8e956a4ea13748260969d20f0e141fa1db73